Re-organize library files
