+++ src/editor/library/container.js
@@ -10,6 +10,7 @@ import { RawHTML, useEffect } from '@wordpress/element';
  */
 import Button from '../../components/button';
 import { updateSCOnEditor } from '../../extensions/style-cards';
+import imageUploader from './uploader';
 
 /**
  * External dependencies
@@ -26,69 +27,6 @@ import {
 	Stats,
 } from 'react-instantsearch-dom';
 import { uniq, isEmpty } from 'lodash';
-
-const placeholderImage = async () => {
-	const ajaxurl = wp.ajax.settings.url;
-	try {
-		const response = await fetch(
-			`${
-				window.location.origin + ajaxurl
-			}?action=maxi_upload_placeholder_image`
-		);
-		const data = await response.json();
-		if (data.error === '404') {
-			console.warn(
-				__(
-					"Can't upload the placeholder image, check directory's permissions",
-					'maxi-blocks'
-				)
-			);
-			return null;
-		}
-		return data;
-	} catch (err) {
-		console.error(
-			__(`Error uploading the placeholder image: ${err}`, 'maxi-blocks')
-		);
-	}
-	return null;
-};
-
-const imageUploader = async imageSrc => {
-	const ajaxurl = wp.ajax.settings.url;
-	try {
-		const response = await fetch(
-			`${
-				window.location.origin + ajaxurl
-			}?action=maxi_upload_pattern_image&maxi_image_to_upload=${imageSrc}`
-		);
-
-		if (!response.ok) {
-			console.warn(
-				__(
-					'The Cloud server is down, using the placeholder image',
-					'maxi-blocks'
-				)
-			);
-			return placeholderImage();
-		}
-
-		const data = await response.json();
-		if (data.error === '404') {
-			console.warn(
-				__(
-					'The original image not found (404) on the Cloud Site, using the placeholder image',
-					'maxi-blocks'
-				)
-			);
-			return placeholderImage();
-		}
-		return data;
-	} catch (err) {
-		console.error(__(`Error uploading the image: ${err}`, 'maxi-blocks'));
-	}
-	return null;
-};
 
 const MasonryItem = props => {
 	const {
