+++ src/editor/library/placeholder.js
@@ -1,0 +1,33 @@
+/**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+const placeholderImage = async () => {
+	const ajaxurl = wp.ajax.settings.url;
+	try {
+		const response = await fetch(
+			`${
+				window.location.origin + ajaxurl
+			}?action=maxi_upload_placeholder_image`
+		);
+		const data = await response.json();
+		if (data.error === '404') {
+			console.warn(
+				__(
+					"Can't upload the placeholder image, check directory's permissions",
+					'maxi-blocks'
+				)
+			);
+			return null;
+		}
+		return data;
+	} catch (err) {
+		console.error(
+			__(`Error uploading the placeholder image: ${err}`, 'maxi-blocks')
+		);
+	}
+	return null;
+};
+
+export default placeholderImage;
